pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "jassist"
version = "0.1.0"
description = "Jassist - Voice Diary Assistant"
requires-python = ">=3.8"
dependencies = [
    "google-auth-oauthlib>=1.0.0",
    "google-auth-httplib2>=0.1.0",
    "python-dotenv",
    "psycopg2-binary",
    "openai>=1.14.0",
    "pyyaml",
    "tenacity>=8.0.0",
    "cryptography>=3.4.0",
    "pathlib>=1.0.1",
    "typing-extensions>=3.7.4",
    "google-api-python-client>=2.86.0",
    "google-auth>=2.19.0",
    "tqdm>=4.65.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-mock>=3.10.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.19.0",
    "pytest-timeout>=2.0.0",
    "pytest-rerunfailures>=10.2",
]

[tool.hatch.build.targets.wheel]
packages = ["jassist"] 